backend/requirements.txt
django-cors-headers==3.7.0
django-filter==2.4.0
djoser==2.1.0
drf-yasg==1.20.0
pip-chill==1.0.1
